Enum tests now run fine on mono thanks to this workaround for a TypeBuilder.SetParent related bug
git-svn-id: https://svn.codehaus.org/boo/trunk@739 2c1201b4-01cd-e047-a400-b836ae1fbc61

+++ src/Boo.Lang.Compiler/TypeSystem/InternalEnumMember.cs
@@ -106,7 +106,7 @@ namespace Boo.Lang.Compiler.TypeSystem
 		{
 			get
 			{
-				return _member.Initializer.Value;
+				return (int)_member.Initializer.Value;
 			}
 		}
